Roles

+++ restapi/products/ProductController.ts
@@ -1,5 +1,6 @@
 import { RequestHandler } from "express";
 import path from "path";
+import { userModel } from "../users/User";
 import { verifyToken } from "../utils/generateToken";
 import { productModel } from "./Product";
 
@@ -45,15 +46,25 @@ export const createProduct: RequestHandler = async (req, res) => {
 };
 
 export const deleteProduct: RequestHandler = async (req, res) => {
-  try {
-    const productFound = await productModel.deleteOne({
-      code: req.params.code,
-    });
-    if (productFound) {
-      return res.json(productFound);
+  const isVerified = verifyToken(
+    req.headers.token + "",
+    req.headers.email + ""
+  );
+  const user = await userModel.findOne({ email: req.headers.email });
+  console.log(user)
+  if (isVerified && user.role === "admin") {
+    try {
+      const productFound = await productModel.deleteOne({
+        code: req.params.code,
+      });
+      if (productFound) {
+        return res.json(productFound);
+      }
+    } catch (error) {
+      res.status(301).json({ message: "The operation didn't succed " + error });
     }
-  } catch (error) {
-    res.status(301).json({ message: "The operation didn't succed " + error });
+  } else {
+    res.status(203).json();
   }
 };
 
@@ -62,7 +73,8 @@ export const updateProduct: RequestHandler = async (req, res) => {
     req.headers.token + "",
     req.headers.email + ""
   );
-  if (isVerified) {
+  const user = userModel.findOne({ email: req.headers.email });
+  if (isVerified && (user.role === "admin" || user.role === "manager")) {
     try {
       console.log(req.body);
       const product = await productModel.findOneAndUpdate(
@@ -77,14 +89,4 @@ export const updateProduct: RequestHandler = async (req, res) => {
   } else {
     res.status(203).json();
   }
-};
-
-export const uploadPhoto: RequestHandler = async (req, res) => {
-  console.log(req.body);
-  return res.json();
-};
-
-export const getPhoto: RequestHandler = async (req, res) => {
-  const photo = await productModel.findOne({ code: req.params.code });
-  return res.json(photo.imagePath);
 };
